Upgrade to Laravel 5.4 completion
