+++ .env.example
@@ -1,11 +1,11 @@
 ##############################################################################################
-####################### ENSURE THE FOLLOWING SETTINGS ARE SET PROPERLY #######################
+####################### ENSURE THE FOLLOWING SETTINGS ARE SET PROPERLY #######################
 ##############################################################################################
 
 APP_ENV=production
 APP_KEY=
 APP_DEBUG=false
-APP_LOG_LEVEL=debug
+APP_LOG_LEVEL=warning
 APP_URL=http://localhost
 
 DB_CONNECTION=mysql
@@ -30,7 +30,7 @@ CURRENCY=£
 
 
 ##############################################################################################
-################## BELOW ARE THE SETTINGS YOU PROBABLY DO NOT NEED TO TOUCH ##################
+################## BELOW ARE THE SETTINGS YOU PROBABLY DO NOT NEED TO TOUCH ##################
 ##############################################################################################
 CACHE_DRIVER=file
 SESSION_DRIVER=file
