+++ composer.json
@@ -9,8 +9,11 @@
     "type": "project",
     "require": {
         "php": ">=5.6.4",
-        "laravel/framework": "5.3.*",
-        "stripe/stripe-php": "3.*"
+        "laravel/framework": "5.4.*",
+        "stripe/stripe-php": "3.*",
+        "laravel/tinker": "^1.0",
+        "laravel/browser-kit-testing": "^1.0",
+        "laravel/dusk": "^1.0"
     },
     "require-dev": {
         "fzaninotto/faker": "~1.4",
@@ -33,7 +36,8 @@
     },
     "autoload-dev": {
         "classmap": [
-            "tests/TestCase.php"
+            "tests/TestCase.php",
+            "tests/BrowserKitTest.php"
         ]
     },
     "scripts": {
