+++ config/app.php
@@ -176,6 +176,8 @@ return [
         App\Providers\AuthServiceProvider::class,
         App\Providers\EventServiceProvider::class,
         App\Providers\RouteServiceProvider::class,
+
+        Laravel\Tinker\TinkerServiceProvider::class
     ],
 
     /*
