+++ config/database.php
@@ -107,15 +107,23 @@ return [
 
     'redis' => [
 
-        'cluster' => false,
+    'client' => 'predis',
 
-        'default' => [
-            'host'     => env('REDIS_HOST', 'localhost'),
-            'password' => env('REDIS_PASSWORD', null),
-            'port'     => env('REDIS_PORT', 6379),
-            'database' => 0,
-        ],
-
+    'options' => [
+        'cluster' => 'redis',
     ],
+
+    'clusters' => [
+        'default' => [
+            [
+                'host' => env('REDIS_HOST', '127.0.0.1'),
+                'password' => env('REDIS_PASSWORD', null),
+                'port' => env('REDIS_PORT', 6379),
+                'database' => 0,
+            ],
+        ],
+    ],
+
+],
 
 ];
