+++ config/mail.php
@@ -109,4 +109,13 @@ return [
 
     'sendmail' => '/usr/sbin/sendmail -bs',
 
+
+    'markdown' => [
+    'theme' => 'default',
+
+    'paths' => [
+        resource_path('views/vendor/mail'),
+    ],
+],
+
 ];
