+++ tests/MainTest.php
@@ -6,7 +6,7 @@ use Illuminate\Foundation\Testing\WithoutMiddleware;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
-class MainTest extends TestCase
+class MainTest extends BrowserKitTest
 {
     use DatabaseMigrations;
