+++ tests/TestCase.php
@@ -1,25 +1,7 @@
 <?php
-
-abstract class TestCase extends Illuminate\Foundation\Testing\TestCase
+namespace Tests;
+use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
+abstract class TestCase extends BaseTestCase
 {
-    /**
-     * The base URL to use while testing the application.
-     *
-     * @var string
-     */
-    protected $baseUrl = 'http://localhost';
-
-    /**
-     * Creates the application.
-     *
-     * @return \Illuminate\Foundation\Application
-     */
-    public function createApplication()
-    {
-        $app = require __DIR__.'/../bootstrap/app.php';
-
-        $app->make(Illuminate\Contracts\Console\Kernel::class)->bootstrap();
-
-        return $app;
-    }
+    use CreatesApplication;
 }
